Updated the code runner to ensure there are no errors in test
