+++ backend/src/model/question.py
@@ -1,12 +1,11 @@
 # Standard library
-from typing import List, Optional, Sequence, Literal, TYPE_CHECKING
+from typing import List, Optional, Sequence, TYPE_CHECKING
 from uuid import UUID, uuid4
 from enum import Enum
 
 # Third-party libraries
 from pydantic import BaseModel, ConfigDict, Field
 from sqlmodel import Field as SQLField, Relationship, SQLModel
-from sqlalchemy import text
 from .question_ownership import QuestionOwnership
 
 from typing import Optional, Sequence
